# api.yaml
openapi: 3.0.0
info:
  title: Temperature Management API
  version: 1.0.0
paths:
...
  /configuration/{id}:
    delete:
      summary: Удаляет конфигурацию
      parameters:
        - name: id
          in: path
          required: true
          description: ID конфигурации
          schema:
            type: integer
      responses:
        '200':
          description: No content
  /configuration/send/{id}:
    post:
      summary: Отправляет конфигурацию в приборы
      parameters:
        - name: id
          in: path
          required: true
          description: ID конфигурации
          schema:
            type: integer
      responses:
        '200':
          description: No content
...
  /configuration/{configName}:
    get:
      summary: Посмотреть конкретную конфигурацию
      parameters:
        - name: configName
          in: path
          required: true
          description: Название конфигурации
          schema:
            type: string
      responses:
        '200':
          description: Detailed configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        device:
                          type: object
                          properties:
                            id:
                              type: integer
                            name:
                              type: string
                            description:
                              type: string
                            serial:
                              type: string
                        roomName:
                          type: string
                        isDeviceEnabled:
                          type: boolean
                  configName:
                    type: string
                  targetTemperature:
                    type: number
...